Migrate `viewport-projection.test.js`
